requirements.txt
pycryptodome>=3.14.1
pytest>=7.1.1
PyQT5>=5.15.2
lorem
